packages/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.3",
    "fastapi[standard]>=0.115.14",
    "lxml>=6.0.0",
    "mistralai>=1.10.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.41",
]

[tool.uv]
package = false
